replaced int with float

+++ findLowestNumber.py
@@ -5,9 +5,9 @@
 #Step 1: Ask for user inputs and assign them to values.
 print("Input three (3) numbers and the program will find the lowest number.\n")
 def userNumbers():
-    a = int(input("Input the first number: "))
-    b = int(input("Input the second number: "))
-    c = int(input("Input the third number: "))
+    a = float(input("Input the first number: "))
+    b = float(input("Input the second number: "))
+    c = float(input("Input the third number: "))
     return a, b, c
 a, b, c = userNumbers()
